FIX: Migrate legacy hamburger menu links to sidebar
