--- assets/javascripts/discourse/initializers/add-hamburger-menu-action.js
@@ -1,25 +1,0 @@
-import { withPluginApi } from "discourse/lib/plugin-api";
-
-function initializeHamburgerMenu(api) {
-  api.decorateWidget("hamburger-menu:generalLinks", () => {
-    return {
-      icon: "calendar-day",
-      route: "discourse-post-event-upcoming-events",
-      label: "discourse_calendar.discourse_post_event.upcoming_events.title",
-    };
-  });
-}
-
-export default {
-  name: "add-hamburger-menu-action",
-
-  initialize(container) {
-    const siteSettings = container.lookup("service:site-settings");
-    if (
-      siteSettings.discourse_post_event_enabled &&
-      siteSettings.sidebar_show_upcoming_events
-    ) {
-      withPluginApi("0.8.7", initializeHamburgerMenu);
-    }
-  },
-};
